+++ demoapp/src/main/java/com/mercadolibre/android/andesui/demoapp/feature/SnackbarShowcaseActivity.kt
@@ -2,18 +2,27 @@ package com.mercadolibre.android.andesui.demoapp.feature
 
 import android.content.Context
 import android.os.Bundle
+import android.support.constraint.Group
 import android.support.v4.view.PagerAdapter
 import android.support.v4.view.ViewPager
 import android.support.v7.app.AppCompatActivity
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.ArrayAdapter
+import android.widget.Spinner
+import android.widget.Switch
 import android.widget.Toast
 import com.mercadolibre.android.andesui.button.AndesButton
 import com.mercadolibre.android.andesui.demoapp.PageIndicator
 import com.mercadolibre.android.andesui.demoapp.R
 import com.mercadolibre.android.andesui.snackbar.AndesSnackbar
+import com.mercadolibre.android.andesui.snackbar.action.AndesSnackbarAction
+import com.mercadolibre.android.andesui.snackbar.duration.AndesSnackbarDuration
 import com.mercadolibre.android.andesui.snackbar.type.AndesSnackbarType
+import com.mercadolibre.android.andesui.textfield.AndesTextfield
+import com.mercadolibre.android.andesui.textfield.state.AndesTextfieldState
+
 
 class SnackbarShowcaseActivity : AppCompatActivity() {
 
@@ -35,32 +44,108 @@ class SnackbarShowcaseActivity : AppCompatActivity() {
     }
 
     private fun addDynamicSnackbar(container: View) {
-        bindAndesSpecsButton(container)
-        val test: AndesButton = container.findViewById(R.id.test)
+        val type: Spinner = container.findViewById(R.id.snackbar_type)
+        val duration: Spinner = container.findViewById(R.id.snackbar_duration)
+        val hasAction: Switch = container.findViewById(R.id.snackbar_has_action)
+        val text: AndesTextfield = container.findViewById(R.id.snackbar_text)
+        val textAction: AndesTextfield = container.findViewById(R.id.snackbar_text_action)
+        val clearButton: AndesButton = container.findViewById(R.id.clear_button)
+        val confirmButton: AndesButton = container.findViewById(R.id.change_button)
+        val actionGroup: Group = container.findViewById(R.id.action_group)
 
-        test.setOnClickListener {
-
-            val snackbar = AndesSnackbar(this, container, AndesSnackbarType.ERROR, "Algo es algo", 5000)
-
-//            snackbar.text = "cambiamos"
-//            snackbar.type = AndesSnackbarType.NEUTRAL
-//            snackbar.duration = 4000
-
-            snackbar.show()
-
+        hasAction.setOnCheckedChangeListener { _, isChecked ->
+            if (isChecked) {
+                actionGroup.visibility = View.VISIBLE
+            } else {
+                actionGroup.visibility = View.GONE
+            }
         }
 
-    }
+        val snackbarType: Spinner = container.findViewById(R.id.snackbar_type)
+        ArrayAdapter.createFromResource(
+                this, R.array.snackbar_type_spinner, android.R.layout.simple_spinner_item)
+                .also { adapter ->
+                    adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
+                    snackbarType.adapter = adapter
+                }
 
+        val snackbarDuration: Spinner = container.findViewById(R.id.snackbar_duration)
+        ArrayAdapter.createFromResource(
+                this, R.array.snackbar_duration_spinner, android.R.layout.simple_spinner_item)
+                .also { adapter ->
+                    adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
+                    snackbarDuration.adapter = adapter
+                }
 
-    private fun bindAndesSpecsButton(container: View) {
-//        container.findViewById<AndesButton>(R.id.andesui_demoapp_andes_badge_specs_button).setOnClickListener {
-//            launchSpecs(container.context, AndesSpecs.BADGE)
-//        }
+        clearButton.setOnClickListener {
+            hasAction.isChecked = false
+            actionGroup.visibility = View.GONE
+
+            text.state = AndesTextfieldState.IDLE
+            text.helper = null
+            text.text = null
+
+            textAction.state = AndesTextfieldState.IDLE
+            textAction.helper = null
+            textAction.text = null
+
+            type.setSelection(0)
+            duration.setSelection(0)
+        }
+
+        confirmButton.setOnClickListener {
+            var hasError = false
+            if (text.text.isNullOrEmpty()) {
+                text.state = AndesTextfieldState.ERROR
+                text.helper = getString(R.string.andesui_demoapp_snackbar_error)
+                hasError = true
+            } else {
+                text.state = AndesTextfieldState.IDLE
+                text.helper = null
+            }
+
+            if (hasAction.isChecked && textAction.text.isNullOrEmpty()) {
+                textAction.state = AndesTextfieldState.ERROR
+                textAction.helper = getString(R.string.andesui_demoapp_snackbar_error)
+                hasError = true
+            } else {
+                textAction.state = AndesTextfieldState.IDLE
+                textAction.helper = null
+            }
+
+            if (hasError) {
+                return@setOnClickListener
+            }
+
+            val selectedType = when (type.selectedItem) {
+                "Neutral" -> AndesSnackbarType.NEUTRAL
+                "Success" -> AndesSnackbarType.SUCCESS
+                else -> AndesSnackbarType.ERROR
+            }
+
+            val selectedDuration = when (duration.selectedItem) {
+                "Long" -> AndesSnackbarDuration.LONG
+                "Normal" -> AndesSnackbarDuration.NORMAL
+                else -> AndesSnackbarDuration.SHORT
+            }
+
+            val snackbar = AndesSnackbar(
+                    this,
+                    container,
+                    selectedType,
+                    text.text!!,
+                    selectedDuration
+            )
+            if (hasAction.isChecked) {
+                snackbar.action = AndesSnackbarAction(textAction.text!!, View.OnClickListener {
+                    Toast.makeText(this, "Callback", Toast.LENGTH_SHORT).show()
+                })
+            }
+            snackbar.show()
+        }
     }
 
     class AndesShowcasePagerAdapter(private val context: Context) : PagerAdapter() {
-
         var views: List<View>
 
         init {